Move `ViToTill` logically to `motion.rs`
